+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
-## Next Release
+## 23.1.0
 ### Features Added
 - Multiple Signing Key configurations can be specified in single YAML file using triple-dash `---` separator. 
 [#689](https://github.com/ConsenSys/web3signer/issues/689)
-- Reloading of signing key configuration file (via `/reload` endpoint) will process new or modified configuration files.
+- Reloading of signing key configuration file (via `/reload` endpoint) will process new or modified configuration files. [#689](https://github.com/ConsenSys/web3signer/issues/689)
+- Updated Teku libraries version to 22.12.0
+
+### Bugs Fixed
+- Upgrade netty libraries to address CVE-2022-41881 and CVE-2022-41915
 
 ---
 ## 22.11.0